Rename recipe to match OpenHAB version
Doing this we correct set PV variable that can be use inside recipe
as Package Version. This also would diferentiate two recipes for two
different versions of OpenHAB that eventualy can be provided.

Internally, recipe will not need OH_VERSION variable anymore and can
rely on PV for the same job.

If the recipe is updated for some reason (ie overhide do_install) the
PR variable should be incremented to signal bitbake that recipe was
updated. This mechanism seems more aligned with Yocto practices.

Signed-off-by: Josenivaldo Benito Jr <jrbenito@benito.qsl.br>